🚚 [rename] CoffeeOptionSection => CategorySection으로 파일명 수정
