+++ src/pages/CreateMyCard/components/index.tsx
@@ -1,2 +1,2 @@
-export { default as CoffeeOptionSection } from './CoffeeOptionSection';
-export { default as CoffeeOption } from './CoffeeOptionSection/CoffeeOption';
+export { default as CategorySection } from './CategorySection';
+export { default as Category } from './CategorySection/Category';
